refactor: 분리된 공통 인터페이스를 import해서 사용하도록 변경
