+++ src/components/organisms/SearchGameListSection/SearchGameListSection.tsx
@@ -7,6 +7,7 @@ import SearchGameList, {
 import { generatePageNumbers } from '@/utils/pagination';
 import { NoResultsMessage } from '@/components/atoms/NoResultsMessage/NoResultsMessage';
 import SearchResultCardSkeleton from '@/components/atoms/SearchResultCardSkeleton/SearchResultCardSkeleton';
+import { ToggleGroupValue } from '@/types/toggle';
 import * as S from './SearchGameListSection.style';
 
 interface SearchGameListSectionProps {
@@ -14,9 +15,9 @@ interface SearchGameListSectionProps {
   keyword: string;
   selectedPage: number;
   totalPages: number;
-  sort: { field: string; order: 'asc' | 'desc' };
+  sort: ToggleGroupValue;
   onPageChange: (page: number) => void;
-  onSortChange: (sort: { field: string; order: 'asc' | 'desc' }) => void;
+  onSortChange: (sort: ToggleGroupValue) => void;
   isLoading: boolean;
 }
